<==== Mercurial data ====>
Branch: meta_resources

+++ common/src/nx/streaming/rtp/camera_time_helper.h
@@ -52,7 +52,7 @@ private:
         int64_t linearize(uint32_t rtpTime);
 
     private:
-        int64_t m_prevTime = 0;
+        uint32_t m_prevTime = 0;
         int64_t m_highPart = 0;
     };
